LU7869 server parks (#12)
+ blaise resources in `operations` module
+ update module init
+ added todo file
+ updated readme
