+++ pyblaise/constants.py
@@ -4,4 +4,6 @@ Template directory
 contains the SOAP XML template files for jinja
 """
 
-TEMPLATE_DIR = "templates"
+from os.path import join
+
+TEMPLATE_DIR = join("operations", "templates")
