+++ pyblaise/operations/templates/auth-header.xml.template
