+++ pyblaise/operations/templates/create-role.xml.template
