+++ pyblaise/operations/templates/create-user.xml.template
