+++ pyblaise/operations/templates/file-footer.xml.template
