+++ pyblaise/operations/templates/get-all-server-park-definitions.xml.template
