+++ pyblaise/operations/templates/get-list-of-instruments.xml.template
