+++ pyblaise/operations/templates/get-roles.xml.template
