+++ pyblaise/operations/templates/get-server-version.xml.template
