+++ pyblaise/operations/templates/get-skills.xml.template
