+++ pyblaise/operations/templates/get-version-response.xml.template
