+++ pyblaise/operations/templates/is-interactive-connection-allowed.xml.template
