+++ pyblaise/operations/templates/remove-instrument.xml.template
