+++ pyblaise/operations/templates/report-user-logout.xml.template
