+++ pyblaise/operations/templates/request-token.xml.template
