+++ pyblaise/operations/templates/survey/opn2004.manifest.template
