+++ pyblaise/tests/requests/test_get_auth_token.py
@@ -6,7 +6,7 @@ from collections import namedtuple
 
 from pyblaise import get_auth_token
 
-from pyblaise.definitions import operations
+from pyblaise.operations.definitions import operations
 
 host_infos = [
     {
