+++ setup.py
@@ -21,7 +21,7 @@ def get_git_version():
       https://www.python.org/dev/peps/pep-0440/
     """
     v = (
-        subprocess.check_output(["git", "describe"], encoding="UTF-8")
+        subprocess.check_output(["git", "describe", "--always"], encoding="UTF-8")
         .strip()
         .split("-")
     )
@@ -55,5 +55,10 @@ setuptools.setup(
     ],
     python_requires=">=3.6",
     install_requires=["jinja2", "requests", "requests_mock"],
-    package_data={"pyblaise": ["templates/*.template", "templates/survey/*.template"]},
+    package_data={
+        "pyblaise": [
+            "operations/templates/*.template",
+            "operations/templates/survey/*.template",
+        ]
+    },
 )
